fix: attrs -> attributes, items -> members, added a failing test that exposes problem with unions in pydantic
