+++ src/pydantic_zarr/__init__.py
@@ -1,2 +1,2 @@
 # ruff: noqa
-from pydantic_zarr.core import ArraySpec, GroupSpec, to_zarr, from_zarr
+from pydantic_zarr.v2 import ArraySpec, GroupSpec, to_zarr, from_zarr
